A new way to define native procedures
More in line with the coming sweeping changes
to evaluation code (more or less a complete
rewrite) to allow for continuations. As such,
this is eventually going to be THE way to
define native procedures. The string business
is not very important and can be optimized
away later. Also good for debugging to have
that information in there, even though we
shouldn't be using it for lookups!